round down ADR

+++ CsStat.Web/Models/Matches/MatchDetailsSquadPlayer.cs
@@ -17,7 +17,7 @@ namespace CsStat.Web.Models.Matches
         public string Kad => $"{Kills}/{Assists}/{Deaths}"; // 10/2/5
         public double KdDiff => Kills - Deaths; // -2
         public double Kd => Math.Round(Deaths > 0 ? (double)Kills / (double)Deaths : Kills, 2); // 1.24
-        public double Adr { get; set; } // 118 ??
-        public double Ud { get; set; } // 24 ??
+        public int Adr { get; set; } // 118 ??
+        public int Ud { get; set; } // 24 ??
     }
 }
